requirements.txt
matplotlib
seaborn
pandas
numpy
networkx
wordcloud
pillow
opencv-python
tqdm
torch
torchvision
nltk
rouge-score
rouge
PyYAML
scikit-learn
tensorboard
fastapi[all]
passlib[bcrypt]
python-jose[cryptography]
transformers
sentencepiece